+++ components/landing/side/side-app-settings.tsx
@@ -30,6 +30,8 @@ import { openAIModelConfig } from '@/config/provider/openai.config';
 import { cohereModelConfig } from '@/config/provider/cohere.config';
 import { huggingFaceModelConfig } from '@/config/provider/huggingface.config';
 
+import { serviceProviderList } from '@/config/app/service.config';
+
 const SideAppSettings = ({ user }: { user: User | null }) => {
     const t = useTranslations('landing.side');
 
@@ -769,52 +771,3 @@ const ExtensionCard = ({}: {}) => {
         </>
     );
 };
-
-interface serviceProviderProps {
-    name: string;
-    value: ServiceProviderProps;
-    status: number;
-}
-
-const serviceProviderList: serviceProviderProps[] = [
-    {
-        name: 'Azure',
-        value: 'Azure',
-        status: 1,
-    },
-    {
-        name: 'Claude',
-        value: 'Claude',
-        status: 2,
-    },
-    {
-        name: 'Cohere',
-        value: 'Cohere',
-        status: 1,
-    },
-    {
-        name: 'Hugging Face',
-        value: 'Hugging Face',
-        status: 1,
-    },
-    {
-        name: 'OpenAI',
-        value: 'OpenAI',
-        status: 1,
-    },
-    {
-        name: 'Team',
-        value: 'Team',
-        status: 1,
-    },
-    {
-        name: 'Custom',
-        value: 'Custom',
-        status: 0,
-    },
-    {
-        name: 'Extension',
-        value: 'Extension',
-        status: 0,
-    },
-];
