--- config/plugin.config.ts
@@ -1,4 +1,0 @@
-export const pluginConfig = [
-    { name: 'search', description: 'Search in internet.' },
-    { name: 'fetch', description: 'Fetch website content.' },
-];
